+++ test/accounts.spec.ts
@@ -13,17 +13,16 @@ describe("accounts", () => {
 
       const result = await dnsimple.accounts.listAccounts();
 
-      expect(result).toEqual({
-        data: [
-          {
-            id: 123,
-            email: "john@example.com",
-            plan_identifier: "dnsimple-personal",
-            created_at: "2011-09-11T17:15:58Z",
-            updated_at: "2016-06-03T15:02:26Z",
-          },
-        ],
-      });
+      expect(result.data).toEqual([
+        {
+          id: 123,
+          email: "john@example.com",
+          plan_identifier: "dnsimple-personal",
+          created_at: "2011-09-11T17:15:58Z",
+          updated_at: "2016-06-03T15:02:26Z",
+        },
+      ]);
+      expect(result.rateLimit).toBeDefined();
     });
   });
 });