+++ test/domain_dnssec.spec.ts
@@ -46,7 +46,7 @@ describe("domains", () => {
         domainId
       );
 
-      expect(response).toEqual({});
+      expect(response.rateLimit).toBeDefined();
     });
   });