+++ test/registrar.spec.ts
@@ -311,7 +311,7 @@ describe("registrar", () => {
         domainId
       );
 
-      expect(response).toEqual({});
+      expect(response.rateLimit).toBeDefined();
     });
   });
 });